[generator] add sphere & cube (#6)
* para o NoBoDy ver

* fix imports

* Templates are now a thing (#7)

* [generator] add sphere

* [utils] improve point class

* [generator] add box draft

* [generator] add cone draft

* [generator] improve plane

* [engine] reee

* [engine] add rotation

* para o NoBoDy ver

* fix imports

* [generator] add sphere

* [utils] improve point class

* [generator] add box draft

* [generator] add cone draft

* [generator] improve plane

* [engine] reee

* [engine] add rotation

* [generator] improve code

+++ src/generator/plane.hpp
@@ -1,11 +1,12 @@
 #ifndef PLANE_H
 #define PLANE_H
-#include <vector>
 #include "../utils/point.hpp"
+
+#include <vector>
 
 class Plane {
 private:
-    double side_length;
+    float side_length;
 
 public:
     Plane(int argc, char** args);
